Added core logic for feedback routing fetching
